Move HTMLDialogElement handling out of Element::removedFromAncestor() https://bugs.webkit.org/show_bug.cgi?id=244146
Reviewed by NOBODY (OOPS!).

Move HTMLDialogElement handling out of Element::removedFromAncestor() since
this is a very hot code path. Instead, move the logic to
HTMLDialogElement::removedFromAncestor().

* Source/WebCore/dom/Document.cpp:
(WebCore::Document::addTopLayerElement):
(WebCore::Document::removeTopLayerElement):
(WebCore::Document::activeModalDialog const):
* Source/WebCore/dom/Document.h:
(WebCore::Document::topLayerElements const):
* Source/WebCore/dom/Element.cpp:
(WebCore::Element::removedFromAncestor):
(WebCore::forEachRenderLayer): Deleted.
(WebCore::Element::addToTopLayer): Deleted.
(WebCore::Element::removeFromTopLayer): Deleted.
* Source/WebCore/dom/Element.h:
(WebCore::Element::isInTopLayer const):
* Source/WebCore/html/HTMLDialogElement.cpp:
(WebCore::forEachRenderLayer):
(WebCore::HTMLDialogElement::removedFromAncestor):
(WebCore::HTMLDialogElement::addToTopLayer):
(WebCore::HTMLDialogElement::removeFromTopLayer):
* Source/WebCore/html/HTMLDialogElement.h:
* Source/WebCore/rendering/RenderLayer.cpp:
* Source/WebCore/style/Styleable.cpp:
